+++ Changelog.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 
+- CS-150 - Added redirect to previous page after form "save and close" and "close" and flash message on "save"
 - CS-178 - Added training list
-- CS-215 - fixed remove attachement button
+- CS-215 - fixed remove attachment button
 - CS-215 - Fixed validation message positioning
 - CS-215 - Fixed validation for file uploads, changes scout's 'logo' attribute to 'profile image'
 - CS-215 - Fixed file remove on new record